+++ tests/unit/test_pathutil.py
@@ -41,7 +41,8 @@ from cylc.flow.pathutil import (
     make_workflow_run_tree,
     parse_rm_dirs,
     remove_dir_and_target,
-    remove_dir_or_file
+    remove_dir_or_file,
+    remove_empty_parents,
 )
 
 from .conftest import MonkeyMock
@@ -407,6 +408,55 @@ def test_remove_dir_or_file(tmp_path: Path):
     assert a_dir.exists()
     remove_dir_or_file(a_dir)
     assert a_dir.exists() is False
+
+
+def test_remove_empty_parents(tmp_path: Path):
+    """Test that _remove_empty_parents() doesn't remove parents containing a
+    sibling."""
+    # -- Setup --
+    reg = 'foo/bar/baz/qux'
+    path = tmp_path.joinpath(reg)
+    tmp_path.joinpath('foo/bar/baz').mkdir(parents=True)
+    # Note qux does not exist, but that shouldn't matter
+    sibling_reg = 'foo/darmok'
+    sibling_path = tmp_path.joinpath(sibling_reg)
+    sibling_path.mkdir()
+    # -- Test --
+    remove_empty_parents(path, reg)
+    assert tmp_path.joinpath('foo/bar').exists() is False
+    assert tmp_path.joinpath('foo').exists() is True
+    # Check it skips non-existent dirs, and stops at the right place too
+    tmp_path.joinpath('foo/bar').mkdir()
+    sibling_path.rmdir()
+    remove_empty_parents(path, reg)
+    assert tmp_path.joinpath('foo').exists() is False
+    assert tmp_path.exists() is True
+
+
+@pytest.mark.parametrize(
+    'path, tail, exc_msg',
+    [
+        pytest.param(
+            'meow/foo/darmok', 'foo/darmok', "path must be absolute",
+            id="relative path"
+        ),
+        pytest.param(
+            '/meow/foo/darmok', '/foo/darmok',
+            "tail must not be an absolute path",
+            id="absolute tail"
+        ),
+        pytest.param(
+            '/meow/foo/darmok', 'foo/jalad',
+            "path '/meow/foo/darmok' does not end with 'foo/jalad'",
+            id="tail not in path"
+        )
+    ]
+)
+def test_remove_empty_parents_bad(path: str, tail: str, exc_msg: str):
+    """Test that _remove_empty_parents() fails appropriately with bad args."""
+    with pytest.raises(ValueError) as exc:
+        remove_empty_parents(path, tail)
+    assert exc_msg in str(exc.value)
 
 
 @pytest.mark.parametrize(
